Added the tail method
This calculates the width and height of a given bounding box and picks
the heights (length of tail in theory), this is done for all the tails
then adds these length of tails into a list and gets smallest and
largest out of all. If largest is not equal to smallest then that means
there is a smaller tail. Meaning if ai model gives a wrong result for the
insect name the tail can be used as a back up for clarification (exaclt
length had to be dropped). Returns if tail is present, tail count and if
smaller tail is present