Change password method
Method that takes in request gets the body of request. Username to
change password for along with the old, new and confirm
password. Checks if old and new password equal if so reply with cant use
old password and. Checks if new and confirm password match if not then
throw up error that passwords dont match. User is authenticated prior to
checking password for security reasons. User password is then updated